support for rules added directly to edit distance function factory
