+++ wordstats/edit_distance.py
@@ -5,11 +5,11 @@ from .edit_distance_function_factory import WordDistanceFactory
 
 class WordDistanceEdit(WordDistanceFactory):
     def __init__(self):
+        super().__init__()
         self.replace_distance = 1
         self.add_distance = 1
-        self.threshold = 0.3
 
-    def edit_distance(self, word1: str, word2: str):
+    def edit_distance_function(self, word1: str, word2: str):
 
         wordLongest, wordShortest = (word1, word2) if len(word1) >= len(word2) else (word2, word1)
 
@@ -21,7 +21,7 @@ class WordDistanceEdit(WordDistanceFactory):
         for i in range(len(wordShortest), len(wordLongest)):
             distance += self.add_distance
 
-        return distance / len(wordLongest) < self.threshold
+        return distance / len(wordLongest)
 
 
     def initialize_from_config(self, config):
